user service

+++ src/main/java/com/wora/gotYou/repositories/DonationRepository.java
@@ -3,5 +3,5 @@ package com.wora.gotYou.repositories;
 import com.wora.gotYou.entities.Donation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DonationRepository extends JpaRepository<Donation, Integer> {
+public interface DonationRepository extends JpaRepository<Donation, Long> {
 }
