+++ src/main/java/com/wora/gotYou/repositories/DonorRepository.java
@@ -3,5 +3,5 @@ package com.wora.gotYou.repositories;
 import com.wora.gotYou.entities.Donor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DonorRepository extends JpaRepository<Donor, Integer> {
+public interface DonorRepository extends JpaRepository<Donor, Long> {
 }
