+++ src/main/java/com/wora/gotYou/repositories/StoriesRepository.java
@@ -3,5 +3,5 @@ package com.wora.gotYou.repositories;
 import com.wora.gotYou.entities.Stories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoriesRepository extends JpaRepository<Stories, Integer> {
+public interface StoriesRepository extends JpaRepository<Stories, Long> {
 }
