+++ src/main/java/com/wora/gotYou/repositories/StudentDonationRepository.java
@@ -3,5 +3,5 @@ package com.wora.gotYou.repositories;
 import com.wora.gotYou.entities.StudentDonation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentDonationRepository extends JpaRepository<StudentDonation, Integer> {
+public interface StudentDonationRepository extends JpaRepository<StudentDonation, Long> {
 }
